*** empty log message ***

+++ include/daqevt.h
@@ -1,4 +1,4 @@
-//  $Id: daqevt.h,v 1.82 2011/05/15 16:16:08 choutko Exp $
+//  $Id: daqevt.h,v 1.82.2.1 2011/06/07 14:15:58 choutko Exp $
 // V. Choutko 15/6/97
 //
 // A.Klimentov June 21, 1997.                   ! add functions
@@ -45,7 +45,7 @@ static unsigned int   huffman_length;
 static bool           _huffman_init;
 static bool           _huffman_swap;
 static void           huffman_init();
-static unsigned short huffman_bit(unsigned short *data, unsigned short i) {
+static unsigned short huffman_bit(unsigned short *data, unsigned int i) {
   return (data[i>>4] & (1 << (0xF - (i&0xF)))) >> (0xF - (i&0xF));
 };
 public:
